List of processes ready
